% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-theme.R
\docType{data}
\name{pal_modality}
\alias{pal_modality}
\title{Modality Colour Palette}
\format{
An object of class \code{character} of length 3.
}
\usage{
pal_modality
}
\description{
Named character vector of 3 colours for imaging modalities.
}
\examples{
pal_modality
}
\keyword{datasets}
