% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-scales.R
\name{scale_artifact}
\alias{scale_artifact}
\alias{scale_colour_artifact}
\alias{scale_fill_artifact}
\title{Colour and Fill Scales for Artifact Grades}
\usage{
scale_colour_artifact(...)

scale_fill_artifact(...)
}
\arguments{
\item{...}{Additional arguments passed to
\code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}} / \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Colour and Fill Scales for Artifact Grades
}
