% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-scales.R
\name{scale_modality}
\alias{scale_modality}
\alias{scale_colour_modality}
\alias{scale_fill_modality}
\title{Colour and Fill Scales for Imaging Modalities}
\usage{
scale_colour_modality(...)

scale_fill_modality(...)
}
\arguments{
\item{...}{Additional arguments passed to
\code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}} / \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Colour and Fill Scales for Imaging Modalities
}
