% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-scales.R
\name{scale_phase}
\alias{scale_phase}
\alias{scale_colour_phase}
\alias{scale_fill_phase}
\title{Colour and Fill Scales for SCI Phases}
\usage{
scale_colour_phase(...)

scale_fill_phase(...)
}
\arguments{
\item{...}{Additional arguments passed to
\code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}} / \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Convenience wrappers around \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}} and
\code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}} using \link{pal_phase}.
}
\examples{
\dontrun{
ggplot(df, aes(x, y, colour = phase)) + geom_point() + scale_colour_phase()
}
}
