% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-theme.R
\docType{data}
\name{pal_artifact}
\alias{pal_artifact}
\title{Artifact Grade Colour Palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
pal_artifact
}
\description{
Named character vector of 5 colours for MRI artifact grades:
none (green), mild (yellow), moderate (orange), severe (red),
non-evaluable (grey).
}
\examples{
pal_artifact
}
\keyword{datasets}
